fix(ui): remove redundant text-gray-600 classes from invoice page
Remove unnecessary "text-gray-600" class from Typography components
in the invoice details page to simplify styling. Also, remove size="sm"
prop from Chip components to standardize their appearance. These changes
improve code clarity and maintain consistent UI styling across the page.
